Downloads/bot-bsjp/requirements.txt
python-telegram-bot==20.6
APScheduler==3.10.4
requests==2.31.0
SQLAlchemy==2.0.25
python-dotenv==1.0.0

yfinance
pandas_ta
